move typeguards to types to prevent dependency cycle
